requirements.txt
discord.py>=2.5.1
fastapi>=0.95.0
uvicorn>=0.23.0
requests>=2.31.0
replit>=7.1.0
discord-py
fastapi
replit
requests
uvicorn
aiohttp>=3.8.0